+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/middleware/OperationMiddleware.kt
@@ -3,7 +3,6 @@ package software.amazon.smithy.swift.codegen.middleware
 import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.model.shapes.ServiceShape
 import software.amazon.smithy.swift.codegen.SwiftWriter
-import software.amazon.smithy.swift.codegen.integration.OperationMiddlewareRenderable
 import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
 import software.amazon.smithy.utils.CodeWriter
 
@@ -11,15 +10,15 @@ import software.amazon.smithy.utils.CodeWriter
  * TODO: We need to solidify an interface to allow rendable middleware to be overridden
  *       This is a placeholder while we figure out how to expose this
  */
-typealias RenderMiddlewareCallback = (CodeWriter, OperationMiddlewareRenderable) -> Boolean
+typealias RenderMiddlewareCallback = (CodeWriter, MiddlewareRenderable) -> Boolean
 
-fun defaultToTrue(codeWriter: CodeWriter, operationMiddlewareRenderable: OperationMiddlewareRenderable): Boolean {
+fun defaultToTrue(codeWriter: CodeWriter, middlewareRenderable: MiddlewareRenderable): Boolean {
     return true
 }
 
 interface OperationMiddleware {
-    fun appendMiddleware(operation: OperationShape, renderableMiddleware: OperationMiddlewareRenderable)
-    fun prependMiddleware(operation: OperationShape, renderableMiddleware: OperationMiddlewareRenderable)
+    fun appendMiddleware(operation: OperationShape, renderableMiddleware: MiddlewareRenderable)
+    fun prependMiddleware(operation: OperationShape, renderableMiddleware: MiddlewareRenderable)
     fun removeMiddleware(operation: OperationShape, step: MiddlewareStep, middlewareName: String)
 
     fun renderMiddleware(
